Allow dossier.yml to override database name
As it stands the DATABASE_URL will override anything in dossier.yml.
Some configurations need those values especially in the database name.

For example where the DATABASE_URL might look like this:

    postgres://postgres@postgres:5432/

With a dossier.yml:

    development:
      database: my_database_dev

Would have caused Dossier to break complaining that **Database "/" does
not exist**

Further still with a DATABASE_URL that has the trailing slash removed
dossier will assume the database name is "" (empty string)

This is especially problematic when the database name changes based on
the RAILS_ENV.