Updated to verify! connection before using
We are getting a lot of PQsocket() can't get socket descriptor errors.

From my understanding this means that we are out of connections BUT this is not true. Connections: 59/400

Based on a small amount of samples this is always happening here:
/lib/dossier/adapter/active_record.rb:34

It seems to me that we are getting a connection but we are then loosing
the connection before it is being used.

By adding verify! we check if the connection is active? and if not we
reset it.