Cache formatted rows
When generating the result array, each iteration through the array would
call the formatting method on each row.

So if you were to, say, display the total number of records at the top
and bottom of the report, then you'd call the formatting methods 3
times: one each time you call size (because it's implemented through
Enumerable, and so depends on each), and once when you're actually
iterating through the rows.

This caches the formatted rows so that multiple invocations within a
same Result instance will never call `format` more than once per row.